Buttons, logging user id in the teams api
I brought back the buttons for navigation until we get the actual
navigation worked out a little better. I also added logging to the API
init and passed it down to the teams api to prove passportjs is
receiving the user cookie and deserializing it for us in the incoming
request.